// asoif/src/app/types/houseTypes.ts
export interface HouseData {
  url: string;
  name: string;
  region: string;
  coatOfArms: string;
  words: string;
  titles: string[];
  seats: string[];
  currentLord: string;
  heir: string;
  overlord: string;
  founded: string;
  founder: string;
  diedOut: string;
  ancestralWeapons: string[];
  cadetBranches: string[];
  swornMembers: [];
}

export interface House {
  houseName: string;
  swornMembers: SwornHouseMember[];
}

export interface SwornHouseMember {
  url: string;
  name: string;
  gender: string;
  culture: string;
  born: string;
  died: string;
  titles: string[];
  aliases: string[];
  father: string;
  mother: string;
  spouse: string;
  allegiances: string[];
  books: string[];
  povBooks: never[];
  tvSeries: string[];
  playedBy: string[];
  houseName: string;
}